# Implement the `count` optimization for the fallback implementation with SWAR