--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -302,3 +302,12 @@ jobs:
     - name: Check formatting
       run: |
         cargo fmt --all -- --check
+  kani:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout repository
+      uses: actions/checkout@v4
+    - name: Run Kani
+      uses: model-checking/kani-github-action@v1
+      with:
+        kani-version: '0.63.0'