… benchmarks/definitions/memchr/sherlock/common.toml
… benchmarks/definitions/memchr/sherlock/never.toml
… benchmarks/definitions/memchr/sherlock/rare.toml
… benchmarks/definitions/memchr/sherlock/uncommon.toml
… benchmarks/definitions/memchr/sherlock/verycommon.toml