Adds the fallback optimized count implementation to rebar engines to compare.

This also includes a [`kani`](https://model-checking.github.io/kani/tool-comparison.html) model checking harness to verify some of the SWAR functions (and a step to run `cargo kani` in CI). Not sure if you'll want to pick it up or not, but it was nice to have confidence that it works for all values.

<details>
<summary>rebar results</summary>

Results from my m1 macbook pro on macos

```
benchmark                          engine                                 /tmp/old.csv      /tmp/new.csv
---------                          ------                                 ------------      ------------
memchr/sherlock/common/huge1       rust/memchr/memchr/fallback/onlycount  92.33us (3.98x)   23.21us (1.00x)
memchr/sherlock/common/small1      rust/memchr/memchr/fallback/onlycount  125.00ns (3.05x)  41.00ns (1.00x)
memchr/sherlock/common/tiny1       rust/memchr/memchr/fallback/onlycount  1.00ns (1.00x)    1.00ns (1.00x)
memchr/sherlock/never/huge1        rust/memchr/memchr/fallback/onlycount  92.33us (3.98x)   23.21us (1.00x)
memchr/sherlock/never/small1       rust/memchr/memchr/fallback/onlycount  125.00ns (3.05x)  41.00ns (1.00x)
memchr/sherlock/never/tiny1        rust/memchr/memchr/fallback/onlycount  1.00ns (1.00x)    1.00ns (1.00x)
memchr/sherlock/never/empty1       rust/memchr/memchr/fallback/onlycount  1.00ns (1.00x)    1.00ns (1.00x)
memchr/sherlock/rare/huge1         rust/memchr/memchr/fallback/onlycount  92.38us (3.98x)   23.21us (1.00x)
memchr/sherlock/rare/small1        rust/memchr/memchr/fallback/onlycount  125.00ns (3.05x)  41.00ns (1.00x)
memchr/sherlock/rare/tiny1         rust/memchr/memchr/fallback/onlycount  1.00ns (1.00x)    1.00ns (1.00x)
memchr/sherlock/uncommon/huge1     rust/memchr/memchr/fallback/onlycount  92.29us (3.98x)   23.21us (1.00x)
memchr/sherlock/uncommon/small1    rust/memchr/memchr/fallback/onlycount  125.00ns (3.05x)  41.00ns (1.00x)
memchr/sherlock/uncommon/tiny1     rust/memchr/memchr/fallback/onlycount  1.00ns (1.00x)    1.00ns (1.00x)
memchr/sherlock/verycommon/huge1   rust/memchr/memchr/fallback/onlycount  92.33us (3.98x)   23.21us (1.00x)
memchr/sherlock/verycommon/small1  rust/memchr/memchr/fallback/onlycount  125.00ns (3.05x)  41.00ns (1.00x)
```

The fallback implementation is still definitely slower than the simd version (of course), but it's closer:

```
benchmark                          rust/memchr/memchr/fallback/onlycount  rust/memchr/memchr/onlycount
---------                          -------------------------------------  ----------------------------
memchr/sherlock/common/huge1       23.21us (1.77x)                        13.08us (1.00x)
memchr/sherlock/common/small1      41.00ns (1.00x)                        41.00ns (1.00x)
memchr/sherlock/common/tiny1       1.00ns (1.00x)                         1.00ns (1.00x)
memchr/sherlock/never/huge1        23.21us (1.77x)                        13.12us (1.00x)
memchr/sherlock/never/small1       41.00ns (1.00x)                        41.00ns (1.00x)
memchr/sherlock/never/tiny1        1.00ns (1.00x)                         1.00ns (1.00x)
memchr/sherlock/never/empty1       1.00ns (1.00x)                         1.00ns (1.00x)
memchr/sherlock/rare/huge1         23.21us (1.77x)                        13.08us (1.00x)
memchr/sherlock/rare/small1        41.00ns (1.00x)                        41.00ns (1.00x)
memchr/sherlock/rare/tiny1         1.00ns (1.00x)                         1.00ns (1.00x)
memchr/sherlock/uncommon/huge1     23.21us (1.77x)                        13.08us (1.00x)
memchr/sherlock/uncommon/small1    41.00ns (1.00x)                        41.00ns (1.00x)
memchr/sherlock/uncommon/tiny1     1.00ns (1.00x)                         1.00ns (1.00x)
memchr/sherlock/verycommon/huge1   23.21us (1.77x)                        13.08us (1.00x)
memchr/sherlock/verycommon/small1  41.00ns (1.00x)                        41.00ns (1.00x)
```

</details>