@Dr-Emann 2025-11-18
@BurntSushi thoughts on merging this?